# packages/web-core/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Duck Engine Web Core API",
    "version": "1.0.0",
    "description": "Core web API for Duck Engine (resource-first materials + file assets; scenes/editor next)",
    "contact": {
      "name": "SpaceDucks Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/admin/resources/{resourceId}/versions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List versions for a resource",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of versions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "count"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResourceVersionWithBindings"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a new version",
        "description": "Supports either a ZIP upload (version.json + slot files) or multipart upload (componentData + any file slot fields). The newly created version becomes the active version.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "description": "If `zip` is provided, it takes precedence over the legacy fields.",
                "properties": {
                  "zip": {
                    "type": "string",
                    "format": "binary",
                    "description": "ZIP containing version.json and slot files (<slot>.<ext>)"
                  },
                  "version": {
                    "type": "integer",
                    "description": "Optional explicit version; if provided must be the next incremental version"
                  },
                  "componentData": {
                    "type": "string",
                    "description": "JSON string (optional)"
                  },
                  "componentType": {
                    "type": "string",
                    "description": "Optional legacy; must match the resource kind when provided"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceVersion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "webcore_auth"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "ResourceVersion": {
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "version",
          "componentType",
          "componentData",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "componentType": {
            "type": "string"
          },
          "componentData": {
            "type": "string",
            "description": "JSON string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FileAsset": {
        "type": "object",
        "required": [
          "id",
          "fileName",
          "contentType",
          "size",
          "sha256",
          "storagePath",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "storagePath": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceBinding": {
        "type": "object",
        "required": [
          "id",
          "resourceVersionId",
          "slot",
          "fileAssetId",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceVersionId": {
            "type": "string"
          },
          "slot": {
            "type": "string"
          },
          "fileAssetId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceBindingWithFile": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResourceBinding"
          },
          {
            "type": "object",
            "required": [
              "fileAsset"
            ],
            "properties": {
              "fileAsset": {
                "$ref": "#/components/schemas/FileAsset"
              }
            }
          }
        ]
      },
      "ResourceVersionWithBindings": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResourceVersion"
          },
          {
            "type": "object",
            "required": [
              "bindings"
            ],
            "properties": {
              "bindings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceBindingWithFile"
                }
              }
            }
          }
        ]
      },
...
    }
  },
...
}